extract createUser test case
